+++ src/terminal/iterm.rs
@@ -113,6 +113,7 @@ pub(crate) const CAPABILITIES: Capabilities = Capabilities {
     window_stacking: false,
     floating_sessions: false,
     capture: true,
+    graphics: false,
 };
 
 pub struct ItermTerminal {